// src/app/core/models/natural-model/address.model.ts
import {CityModel} from "../../../features-master/city/models/city.model";
import {UserModel} from "../../../features-master/users/models/user.model";

export class AddressModel{
  address: any;
  addressSuite: string;
  country: string;
  city: string;
  region: string;
  postalCode: string;
  arrondissement: string;
}